Added main functionalities of new UI
Old code is kept unaltered as much as possible. Albeit, code modularity could be improved and redundancy could be reduced, I did not do that deliberately. It looks like LeetCode is not going to hang on to their old UI for very long and eventually code for old UI will be rendered useless. So, there is no point blending codes of the two cases trying to modularize the code. Instead, I tried to code in a way so that old code could be easily purged (though there are so many repetitive and redundant code) without having to worry about messing up the code for the new UI.